ISO support, but ioapic doesnt works

+++ kernel/src/include/arch/x86_64/interrupts/ioapic.hpp
@@ -7,7 +7,7 @@ class IOAPIC {
 public:
     static void Init();
 
-    static void SetEntry(uint8_t vector,uint8_t irq,uint32_t flags,uint64_t lapic);
+    static void SetEntry(uint8_t vector,uint8_t irq,uint64_t flags,uint64_t lapic);
 
     static void Write(uint64_t phys_base,uint32_t reg,uint32_t value);
     static uint32_t Read(uint64_t phys_base,uint32_t reg);
